refactor: Remove babel/register for instant production server startup
