+++ .gitignore
@@ -1,4 +1,5 @@
 dist
+build
 node_modules/*
 server/scripts
 .env
