+++ Dockerfile
@@ -1,8 +1,9 @@
-FROM node:12-alpine
+FROM node:14-alpine
 
 ENV PATH /opt/outline/node_modules/.bin:/opt/node_modules/.bin:$PATH
 ENV NODE_PATH /opt/outline/node_modules:/opt/node_modules
 ENV APP_PATH /opt/outline
+ENV NODE_ENV production
 RUN mkdir -p $APP_PATH
 
 WORKDIR $APP_PATH
