--- init.js
@@ -1,4 +1,0 @@
-// @flow
-require("@babel/register");
-require("@babel/polyfill");
-require("dotenv").config({ silent: true });
