+++ package.json
@@ -1,14 +1,16 @@
 {
   "name": "outline",
   "private": true,
+  "license": "Business Source License 1.1",
   "main": "index.js",
   "scripts": {
-    "clean": "rimraf dist",
-    "build:webpack": "NODE_ENV=production webpack --config webpack.config.prod.js",
-    "build:analyze": "NODE_ENV=production webpack --config webpack.config.prod.js --json > stats.json",
-    "build": "yarn clean && yarn build:webpack",
-    "start": "NODE_ENV=production node index.js",
-    "dev": "NODE_ENV=development nodemon --watch server index.js",
+    "clean": "rimraf dist && rimraf build",
+    "build:server": "babel -d ./build/server ./server && babel -d ./build/shared ./shared && ln -sf \"$(pwd)/package.json\" ./build && ln -sf \"$(pwd)/webpack.config.dev.js\" ./build",
+    "build:webpack": "webpack --config webpack.config.prod.js",
+    "build:analyze": "webpack --config webpack.config.prod.js --json > stats.json",
+    "build": "yarn clean && yarn build:webpack && yarn build:server",
+    "start": "node ./build/server/index.js",
+    "dev": "nodemon --exec \"yarn build:server && node build/server/index.js\" -e js --ignore build/",
     "lint": "eslint app server shared",
     "flow": "flow",
     "deploy": "git push heroku master",
@@ -64,7 +66,6 @@
     "@babel/preset-env": "^7.11.0",
     "@babel/preset-flow": "^7.10.4",
     "@babel/preset-react": "^7.10.4",
-    "@babel/register": "^7.10.5",
     "@rehooks/window-scroll-position": "^1.0.1",
     "@sentry/node": "^5.12.2",
     "@tippy.js/react": "^2.2.2",
@@ -159,6 +160,7 @@
     "validator": "5.2.0"
   },
   "devDependencies": {
+    "@babel/cli": "^7.10.5",
     "@relative-ci/agent": "^1.3.0",
     "babel-eslint": "^10.1.0",
     "babel-jest": "^26.2.2",
