+++ webpack.config.js
@@ -9,7 +9,7 @@ require('dotenv').config({ silent: true });
 
 module.exports = {
   output: {
-    path: path.join(__dirname, 'dist'),
+    path: path.join(__dirname, 'build/app'),
     filename: '[name].[hash].js',
     publicPath: '/static/',
   },
